+++ src/exactly_lib/impls/instructions/setup/env.py
@@ -5,6 +5,6 @@ from exactly_lib.impls.instructions.setup.utils import instruction_from_parts
 
 def setup(instruction_name: str) -> SingleInstructionSetup:
     return SingleInstructionSetup(
-        instruction_from_parts.Parser(parse.PARTS_PARSER),
+        instruction_from_parts.Parser(parse.parts_parser(phase_is_after_act=False)),
         doc.TheInstructionDocumentation(instruction_name,
                                         is_in_setup_phase=True))